Bless this wonderful Emacs config
This repository is supposed to hold my personal Emacs configuration. I
want to avoid running into Emacs bankruptcy. I also want to avoid
spending a lot of time configuring Emacs. The new readme file outlines
the principles this config will be built on. This also is a
personalization of Purcell's config.